refactor(visualization): consolidate hooks into useVisualization.ts, delete visualization-api.ts
